Add User app with models, views, and forms
This commit introduces the User module to the current application. It include models `User` and `TelegramUser`, views for user login, registration and profile, and forms for login and registration. It also modifies some fields of the User model (such as `phone_number` and other data fields) and adds relevant migrations. Additionally, it provides Django admin configuration for the User model.