Remove ngrok.yml configuration file
The ngrok.yml configuration file was removed from the project. This is part of the process to simplify the project setup and focus on creating Docker images, which is reflected in the updated Dockerfile and Docker compose files. The removal of this service doesn't affect the main functionality of the project.