Remove fullscreen layout and logout link from user page
The fullscreen layout set by the "vh-100" class has been removed for a more standard layout in the user template. In addition, the direct logout link from the user page has been deleted to move it to a more suitable location in the application. These layout and functionality changes will provide an improved user experience.