Relocate docker-entrypoint script and update Dockerfile
The docker-entrypoint script has been moved from the "entrypoints" directory to the project root. As a result, the Dockerfile and compose.yaml file have been updated to reference the new script location. Additionally, the command to start the Gunicorn server has been shifted from the entrypoint script to the Dockerfile.