Refactor TelegramUser classes and paths in user app
The TelegramUser related views and corresponding urls in the user application are refactored for better clarity. The class names now explicitly state their behavior (List, Detail, Create, Update, Delete) with addition of View at the end, making the code more readable and easier to understand. Same changes are reflected in the urls.py file as well.