Update Dockerfile and compose.yaml, add DockerfileBot and new requirements
The Dockerfile and compose.yaml were updated to reflect the new structure of the bot application. A separate Dockerfile for the bot (DockerfileBot) has been created. The requirements.txt file now includes new dependencies necessary for the bot functionality. This paves the way for the separation of concerns and ease of maintenance in the future.