Remove webhook methods from telegram_bot model
The `set_webhook` and `delete_webhook` methods were removed from the telegram_bot model file. This change was necessary due to restructuring of the application, and these methods and their functionality will now be handled separately in another component of the application. The removal also helps clean up the model file and adhere to the single responsibility principle.