Add language setting functionality to the Telegram bot
Implemented a feature that allows Telegram bot users to change the display language in the bot. The command '/start' now prompts the user to choose a language if it's a new user or if the user doesn't have a language setting. This feature improves the Telegram bot's usability and user-friendliness by providing a tailored user experience.